+++ tests/micropython_spi_run.py
@@ -8,8 +8,8 @@ import argparse
 import os
 import time
 
-from rp2040py.cli.bootrom import BOOTROM_B1
-from rp2040py.cli.load_flash import load_micropython_flash_image, load_uf2
+from rp2040py.device.bootrom import BOOTROM_B1
+from rp2040py.device.load_flash import load_micropython_flash_image, load_uf2
 from rp2040py.gpio_pin import GPIOPinState
 from rp2040py.simulator import Simulator
 from rp2040py.utils.logging import ConsoleLogger, LogLevel